requirements.txt
ultralytics[export]
opencv-python
python-dotenv
tensorflow
ml-dtypes
mediapipe
picamera2